requirements.txt
asgiref==3.3.1
Babel==2.9.0
dj-database-url==0.5.0
Django==3.0.11
django-extra-views==0.13.0
django-haystack==3.0
django-heroku==0.3.1
django-oscar==2.1
django-oscar-api==2.1.1
django-phonenumber-field==3.0.1
django-rest-framework==0.1.0
django-tables2==2.2.1
django-treebeard==4.4
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
factory-boy==2.12.0
Faker==5.6.5
gunicorn==20.0.4
phonenumbers==8.12.16
Pillow==8.0.1
psycopg2==2.8.6
purl==1.5
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
pyyaml==5.4.1
